refactor: split MemberController into dedicated ViewModels
- AuthViewModel: login, register, init, testLogin
- MemberViewModel: member state, refresh, image upload
- EmailViewModel: associateEmail, forgotIdentifier, validateAction
- FollowViewModel: follow/unfollow animes and episodes
- ConflictEmailException: extracted to its own file
- MemberController: now a facade delegating to the new ViewModels
- Tests: updated to use AuthViewModel directly
- dart analyze: 0 issues, flutter test: 48/48
