+++ lib/ui/viewmodels/conflict_email_exception.dart
@@ -1,0 +1,3 @@
+class ConflictEmailException implements Exception {
+  const ConflictEmailException();
+}
